# bsn-ui/src/app/services/models/book-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BookResponse {
  id?: number;
  title?: string;
  author?: string;
  isbn?: string;
  synopsis?: string;
  owner?: string;
  rate?: number;
  cover?: Array<string>;
  archived?: boolean;
  shareable?: boolean;
}