# bsn-ui/src/app/services/models/registration-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface RegistrationRequest {
  email: string;
  firstName: string;
  lastName: string;
  password: string;
}