# bsn/src/main/java/com/bassem/bsn/auth/AuthenticationResponse.java
package com.bassem.bsn.auth;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class AuthenticationResponse {
    private String token;
}