# bsn/src/main/java/com/bassem/bsn/book/BorrowedBookResponse.java
package com.bassem.bsn.book;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class BorrowedBookResponse {
    private Integer id;
    private String title;
    private String author;
    private String isbn;
    private Double rate;
    private boolean returned;
    private boolean returnedApproved;
}